[SQL] Add support for TableMacro UDF
