+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/impl/BeamCalciteTable.java
@@ -17,6 +17,7 @@
  */
 package org.apache.beam.sdk.extensions.sql.impl;
 
+import com.google.common.collect.ImmutableMap;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -40,15 +41,19 @@ import org.apache.calcite.schema.SchemaPlus;
 import org.apache.calcite.schema.TranslatableTable;
 
 /** Adapter from {@link BeamSqlTable} to a calcite Table. */
-class BeamCalciteTable extends AbstractQueryableTable
+public class BeamCalciteTable extends AbstractQueryableTable
     implements ModifiableTable, TranslatableTable {
   private final BeamSqlTable beamTable;
   private final Map<String, String> pipelineOptions;
 
-  public BeamCalciteTable(BeamSqlTable beamTable, Map<String, String> pipelineOptions) {
+  BeamCalciteTable(BeamSqlTable beamTable, Map<String, String> pipelineOptions) {
     super(Object[].class);
     this.beamTable = beamTable;
     this.pipelineOptions = pipelineOptions;
+  }
+
+  public static BeamCalciteTable of(BeamSqlTable table) {
+    return new BeamCalciteTable(table, ImmutableMap.of());
   }
 
   @Override
