+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/impl/UdfImpl.java
@@ -17,64 +17,26 @@
  */
 package org.apache.beam.sdk.extensions.sql.impl;
 
-import static org.apache.calcite.util.Static.RESOURCE;
-
-import com.google.common.collect.ImmutableMultimap;
-import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import org.apache.beam.sdk.extensions.sql.impl.utils.CalciteUtils;
-import org.apache.calcite.adapter.enumerable.CallImplementor;
-import org.apache.calcite.adapter.enumerable.NullPolicy;
-import org.apache.calcite.adapter.enumerable.ReflectiveCallNotNullImplementor;
-import org.apache.calcite.adapter.enumerable.RexImpTable;
-import org.apache.calcite.linq4j.function.SemiStrict;
-import org.apache.calcite.linq4j.function.Strict;
-import org.apache.calcite.rel.type.RelDataType;
-import org.apache.calcite.rel.type.RelDataTypeFactory;
-import org.apache.calcite.schema.ImplementableFunction;
-import org.apache.calcite.schema.ScalarFunction;
-import org.apache.calcite.schema.impl.ScalarFunctionImpl;
-import org.apache.calcite.sql.SqlOperatorBinding;
+import org.apache.calcite.schema.Function;
+import org.apache.calcite.schema.TranslatableTable;
+import org.apache.calcite.schema.impl.TableMacroImpl;
 
-/** Beam-customized version from {@link ScalarFunctionImpl}, to address BEAM-5921. */
-public class UdfImpl extends UdfImplReflectiveFunctionBase
-    implements ScalarFunction, ImplementableFunction {
+/** Beam-customized facade behind {@link Function} to address BEAM-5921. */
+class UdfImpl {
 
-  private final CallImplementor implementor;
-
-  /** Private constructor. */
-  private UdfImpl(Method method, CallImplementor implementor) {
-    super(method);
-    this.implementor = implementor;
-  }
-
-  /** Creates {@link org.apache.calcite.schema.ScalarFunction} for each method in a given class. */
-  public static ImmutableMultimap<String, ScalarFunction> createAll(Class<?> clazz) {
-    final ImmutableMultimap.Builder<String, ScalarFunction> builder = ImmutableMultimap.builder();
-    for (Method method : clazz.getMethods()) {
-      if (method.getDeclaringClass() == Object.class) {
-        continue;
-      }
-      if (!Modifier.isStatic(method.getModifiers()) && !classHasPublicZeroArgsConstructor(clazz)) {
-        continue;
-      }
-      final ScalarFunction function = create(method);
-      builder.put(method.getName(), function);
-    }
-    return builder.build();
-  }
+  private UdfImpl() {}
 
   /**
-   * Creates {@link org.apache.calcite.schema.ScalarFunction} from given class.
+   * Creates {@link org.apache.calcite.schema.Function} from given class.
    *
    * <p>If a method of the given name is not found or it does not suit, returns {@code null}.
    *
    * @param clazz class that is used to implement the function
    * @param methodName Method name (typically "eval")
-   * @return created {@link ScalarFunction} or null
+   * @return created {@link Function} or null
    */
-  public static ScalarFunction create(Class<?> clazz, String methodName) {
+  public static Function create(Class<?> clazz, String methodName) {
     final Method method = findMethod(clazz, methodName);
     if (method == null) {
       return null;
@@ -83,91 +45,17 @@ public class UdfImpl extends UdfImplReflectiveFunctionBase
   }
 
   /**
-   * Creates {@link org.apache.calcite.schema.ScalarFunction} from given method. When {@code eval}
-   * method does not suit, {@code null} is returned.
+   * Creates {@link org.apache.calcite.schema.Function} from given method.
    *
    * @param method method that is used to implement the function
-   * @return created {@link ScalarFunction} or null
+   * @return created {@link Function} or null
    */
-  public static ScalarFunction create(Method method) {
-    if (!Modifier.isStatic(method.getModifiers())) {
-      Class clazz = method.getDeclaringClass();
-      if (!classHasPublicZeroArgsConstructor(clazz)) {
-        throw RESOURCE.requireDefaultConstructor(clazz.getName()).ex();
-      }
-    }
-    if (method.getExceptionTypes().length != 0) {
-      throw new RuntimeException(method.getName() + " must not throw checked exception");
-    }
-    CallImplementor implementor = createImplementor(method);
-    return new UdfImpl(method, implementor);
-  }
-
-  @Override
-  public RelDataType getReturnType(RelDataTypeFactory typeFactory) {
-    return CalciteUtils.sqlTypeWithAutoCast(typeFactory, method.getReturnType());
-  }
-
-  @Override
-  public CallImplementor getImplementor() {
-    return implementor;
-  }
-
-  private static CallImplementor createImplementor(final Method method) {
-    final NullPolicy nullPolicy = getNullPolicy(method);
-    return RexImpTable.createImplementor(
-        new ReflectiveCallNotNullImplementor(method), nullPolicy, false);
-  }
-
-  private static NullPolicy getNullPolicy(Method m) {
-    if (m.getAnnotation(Strict.class) != null) {
-      return NullPolicy.STRICT;
-    } else if (m.getAnnotation(SemiStrict.class) != null) {
-      return NullPolicy.SEMI_STRICT;
-    } else if (m.getDeclaringClass().getAnnotation(Strict.class) != null) {
-      return NullPolicy.STRICT;
-    } else if (m.getDeclaringClass().getAnnotation(SemiStrict.class) != null) {
-      return NullPolicy.SEMI_STRICT;
+  public static Function create(Method method) {
+    if (TranslatableTable.class.isAssignableFrom(method.getReturnType())) {
+      return TableMacroImpl.create(method);
     } else {
-      return NullPolicy.NONE;
+      return ScalarFunctionImpl.create(method);
     }
-  }
-
-  public RelDataType getReturnType(RelDataTypeFactory typeFactory, SqlOperatorBinding opBinding) {
-    // Strict and semi-strict functions can return null even if their Java
-    // functions return a primitive type. Because when one of their arguments
-    // is null, they won't even be called.
-    final RelDataType returnType = getReturnType(typeFactory);
-    switch (getNullPolicy(method)) {
-      case STRICT:
-        for (RelDataType type : opBinding.collectOperandTypes()) {
-          if (type.isNullable()) {
-            return typeFactory.createTypeWithNullability(returnType, true);
-          }
-        }
-        break;
-      case SEMI_STRICT:
-        return typeFactory.createTypeWithNullability(returnType, true);
-      default:
-        break;
-    }
-    return returnType;
-  }
-
-  /**
-   * Verifies if given class has public constructor with zero arguments.
-   *
-   * @param clazz class to verify
-   * @return true if given class has public constructor with zero arguments
-   */
-  static boolean classHasPublicZeroArgsConstructor(Class<?> clazz) {
-    for (Constructor<?> constructor : clazz.getConstructors()) {
-      if (constructor.getParameterTypes().length == 0
-          && Modifier.isPublic(constructor.getModifiers())) {
-        return true;
-      }
-    }
-    return false;
   }
 
   /*
@@ -185,5 +73,3 @@ public class UdfImpl extends UdfImplReflectiveFunctionBase
     return null;
   }
 }
-
-// End ScalarFunctionImpl.java
